backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
twilio==8.10.0
razorpay==1.4.1

# Production dependencies
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
